# frontend/src-gen/api-docs.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "CampusCoffeeConsumption API",
    "description" : "REST API for tracking the coffee consumption of SE@UHD users: users bump their own count via a secret capability link, and admins manage users and adjust counts.",
    "version" : "0.0.0"
  },
...
  "paths" : {
...
    "/api/users/{userId}/consumption" : {
      "get" : {
        "tags" : [ "User consumption (admin)" ],
        "summary" : "Get a user's current coffee total and recent changes.",
        "operationId" : "get",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "description" : "Unique identifier of the user.",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "limit",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "exclusiveMinimum" : 0,
            "maximum" : 100
          }
        }, {
          "name" : "offset",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "maximum" : 1000000,
            "minimum" : 0
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ConsumptionDto"
                }
              }
            }
          },
          "400" : {
            "description" : "The request is malformed or violates a validation rule.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "The request is unauthenticated (a missing, unknown, or rotated credential).",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "The authenticated caller is not permitted to perform this operation.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "A referenced resource does not exist.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409" : {
            "description" : "The request conflicts with the current state of the resource.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "User consumption (admin)" ],
        "summary" : "Override a user's coffee count to an explicit value (admin correction).",
        "operationId" : "overrideTotal",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "description" : "Unique identifier of the user.",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ConsumptionOverrideDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ConsumptionDto"
                }
              }
            }
          },
          "400" : {
            "description" : "The request is malformed or violates a validation rule.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "The request is unauthenticated (a missing, unknown, or rotated credential).",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "The authenticated caller is not permitted to perform this operation.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "A referenced resource does not exist.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409" : {
            "description" : "The request conflicts with the current state of the resource.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "User consumption (admin)" ],
        "summary" : "Apply a +1/-1 change to a user's coffee count.",
        "operationId" : "change",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "description" : "Unique identifier of the user.",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ConsumptionDeltaDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ConsumptionDto"
                }
              }
            }
          },
          "400" : {
            "description" : "The request is malformed or violates a validation rule.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "The request is unauthenticated (a missing, unknown, or rotated credential).",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "The authenticated caller is not permitted to perform this operation.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "A referenced resource does not exist.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409" : {
            "description" : "The request conflicts with the current state of the resource.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/{userId}/consumption/rating" : {
      "put" : {
        "tags" : [ "User consumption (admin)" ],
        "summary" : "Rate the beans of a user's current cup on their behalf (admin).",
        "operationId" : "rate",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "description" : "Unique identifier of the user.",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RatingRequestDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ConsumptionDto"
                }
              }
            }
          },
          "400" : {
            "description" : "The request is malformed or violates a validation rule.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "The request is unauthenticated (a missing, unknown, or rotated credential).",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "The authenticated caller is not permitted to perform this operation.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "A referenced resource does not exist.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409" : {
            "description" : "The request conflicts with the current state of the resource.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/consumption/rating" : {
      "put" : {
        "tags" : [ "Consumption" ],
        "summary" : "Rate the beans of the authenticated user's current coffee (one to five).",
        "operationId" : "rate_1",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RatingRequestDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserSummaryDto"
                }
              }
            }
          },
          "400" : {
            "description" : "The request is malformed or violates a validation rule.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "The request is unauthenticated (a missing, unknown, or rotated credential).",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "The authenticated caller is not permitted to perform this operation.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "A referenced resource does not exist.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409" : {
            "description" : "The request conflicts with the current state of the resource.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/users/{userId}/consumption/cancel" : {
      "post" : {
        "tags" : [ "User consumption (admin)" ],
        "summary" : "Undo a user's most recent coffee within the grace period (admin).",
        "operationId" : "cancel",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "description" : "Unique identifier of the user.",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ConsumptionDto"
                }
              }
            }
          },
          "400" : {
            "description" : "The request is malformed or violates a validation rule.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "The request is unauthenticated (a missing, unknown, or rotated credential).",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "The authenticated caller is not permitted to perform this operation.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "A referenced resource does not exist.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409" : {
            "description" : "The request conflicts with the current state of the resource.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/kitty/adjustment" : {
      "post" : {
        "tags" : [ "Kitty" ],
        "summary" : "Adjust the kitty (initial float or correction).",
        "operationId" : "adjustment",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/AdjustmentRequestDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Created",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PaymentDto"
                }
              }
            }
          },
          "400" : {
            "description" : "The request is malformed or violates a validation rule.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "The request is unauthenticated (a missing, unknown, or rotated credential).",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "The authenticated caller is not permitted to perform this operation.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "A referenced resource does not exist.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409" : {
            "description" : "The request conflicts with the current state of the resource.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/consumption" : {
      "post" : {
        "tags" : [ "Consumption" ],
        "summary" : "Add one coffee to the authenticated user's count.",
        "operationId" : "add",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserSummaryDto"
                }
              }
            }
          },
          "400" : {
            "description" : "The request is malformed or violates a validation rule.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "The request is unauthenticated (a missing, unknown, or rotated credential).",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "The authenticated caller is not permitted to perform this operation.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "A referenced resource does not exist.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409" : {
            "description" : "The request conflicts with the current state of the resource.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/consumption/cancel" : {
      "post" : {
        "tags" : [ "Consumption" ],
        "summary" : "Undo the authenticated user's most recent coffee (within the grace period).",
        "operationId" : "cancel_1",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserSummaryDto"
                }
              }
            }
          },
          "400" : {
            "description" : "The request is malformed or violates a validation rule.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "The request is unauthenticated (a missing, unknown, or rotated credential).",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "The authenticated caller is not permitted to perform this operation.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "A referenced resource does not exist.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409" : {
            "description" : "The request conflicts with the current state of the resource.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/users/{userId}/summary" : {
      "get" : {
        "tags" : [ "Admin accounting" ],
        "summary" : "Get a user's landing summary (admin).",
        "operationId" : "userSummary",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "limit",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "exclusiveMinimum" : 0,
            "maximum" : 100
          }
        }, {
          "name" : "offset",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "maximum" : 1000000,
            "minimum" : 0
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserSummaryDto"
                }
              }
            }
          },
          "400" : {
            "description" : "The request is malformed or violates a validation rule.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "The request is unauthenticated (a missing, unknown, or rotated credential).",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "The authenticated caller is not permitted to perform this operation.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "A referenced resource does not exist.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409" : {
            "description" : "The request conflicts with the current state of the resource.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/users/activity" : {
      "get" : {
        "tags" : [ "Admin activity" ],
        "summary" : "Get a page of the whole-installation activity feed.",
        "operationId" : "activity",
        "parameters" : [ {
          "name" : "limit",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "exclusiveMinimum" : 0,
            "maximum" : 100
          }
        }, {
          "name" : "offset",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "maximum" : 1000000,
            "minimum" : 0
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/GlobalActivityEntryDto"
                  }
                }
              }
            }
          },
          "400" : {
            "description" : "The request is malformed or violates a validation rule.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "The request is unauthenticated (a missing, unknown, or rotated credential).",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "The authenticated caller is not permitted to perform this operation.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "A referenced resource does not exist.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409" : {
            "description" : "The request conflicts with the current state of the resource.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "ConsumptionOverrideDto" : {
        "type" : "object",
        "description" : "The new, non-negative total and an optional note.",
        "properties" : {
          "total" : {
            "type" : "integer",
            "format" : "int32",
            "maximum" : 1000000,
            "minimum" : 0
          },
          "note" : {
            "type" : [ "string", "null" ],
            "maxLength" : 500,
            "minLength" : 0
          }
        },
        "required" : [ "total" ]
      },
...
      "ConsumptionDto" : {
        "type" : "object",
        "properties" : {
          "total" : {
            "type" : "integer",
            "format" : "int32"
          },
          "changes" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ConsumptionChangeDto"
            }
          }
        },
        "required" : [ "changes", "total" ]
      },
      "RatingRequestDto" : {
        "type" : "object",
        "properties" : {
          "beanId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "value" : {
            "type" : "integer",
            "format" : "int32",
            "maximum" : 5,
            "minimum" : 1
          }
        },
        "required" : [ "beanId", "value" ]
      },
...
      "ConsumptionDeltaDto" : {
        "type" : "object",
        "description" : "The single-step delta to apply (+1 or -1).",
        "properties" : {
          "delta" : {
            "type" : "integer",
            "format" : "int32",
            "maximum" : 1,
            "minimum" : -1
          }
        },
        "required" : [ "delta" ]
      },
...
      "AdjustmentRequestDto" : {
        "type" : "object",
        "properties" : {
          "amountCents" : {
            "type" : "integer",
            "format" : "int32",
            "maximum" : 100000,
            "minimum" : -100000
          },
          "note" : {
            "type" : [ "string", "null" ],
            "maxLength" : 500,
            "minLength" : 0
          }
        },
        "required" : [ "amountCents" ]
      },
...
      "ErrorResponse" : {
        "properties" : {
          "errorCode" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          },
          "statusCode" : {
            "type" : "integer",
            "format" : "int32"
          },
          "statusMessage" : {
            "type" : "string"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          },
          "path" : {
            "type" : "string"
          }
        }
      }
    }
  }
}